Chore: Network Layer
Implement the network layer, GET, POST.
