+++ pubspec.yaml
@@ -34,6 +34,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  # package for http calls
+  http: ^0.13.4
 
 dev_dependencies:
   flutter_test:
